+++ src/BetterSafe.hpp
@@ -10,7 +10,7 @@ struct SafeLevel {
     int id;
     int timelyID;
     std::vector<SafeDate> dates;
-    bool weekly;
+    GJTimedLevelType type;
     int tier;
 };
 
@@ -18,10 +18,12 @@ class BetterSafe {
 public:
     inline static std::vector<SafeLevel> DAILY_SAFE;
     inline static std::vector<SafeLevel> WEEKLY_SAFE;
+    inline static std::vector<SafeLevel> EVENT_SAFE;
 
     static SafeDate parseDate(const std::string&);
-    static void loadDailySafe(geode::EventListener<geode::utils::web::WebTask>&&, LoadingCircle*, std::function<void()> const&);
-    static void loadWeeklySafe(geode::EventListener<geode::utils::web::WebTask>&&, LoadingCircle*, std::function<void()> const&);
-    static std::vector<SafeLevel> getMonth(int year, int month, bool weekly);
+    static SafeDate dateFromTime(time_t);
+    static void loadSafe(GJTimedLevelType, geode::EventListener<geode::utils::web::WebTask>&&, LoadingCircle*, std::function<void()> const&);
+    static std::vector<SafeLevel> getMonth(int year, int month, GJTimedLevelType type);
+    static std::vector<SafeLevel>& getSafeLevels(GJTimedLevelType type);
     static int getDifficultyFromLevel(GJGameLevel* level);
 };